# internal/handler/openapi.yaml
openapi: 3.1.0
...
info:
  title: Cronsole API
  version: "2.0.0"
  summary: A central scheduler that calls your services on time.
  description: |
    Two audiences, and the difference decides which key you send.

    **A project** authenticates with `X-API-Key` and reaches its own jobs and
    nothing else. This is the integration path: a project declares its whole
    schedule from its own deploy pipeline with `POST /sync`, and can trigger and
    read its own runs.

    **An operator** authenticates with a bearer token from `POST /login` and
    reaches whatever their roles reach, which for a platform administrator is
    everything. This is the scripting path.

    ### Responses

    Every response is the same envelope:

    ```json
    { "status": true, "data": { }, "message": "", "errors": [ ] }
    ```

    `status` is the plain success flag; `data` carries the payload; `message`
    carries a human readable error; `errors` is present only on a validation
    failure and names every field at once.

    ### Language

    **These messages are always English.** The interface is available in English
    and Turkish, but the language middleware is not mounted on `/api/v1`: an
    `Accept-Language` header changes nothing here. An error whose wording depends
    on a request header is one nobody can grep for in a log.

    ### Scoping

    A list returns what the caller may see. A single row outside that set answers
...
  license:
    name: Apache-2.0
...
  contact:
...
servers:
...
security:
  - projectKey: []
paths:
...
  /jobs/{code}/run:
    post:
      tags: [Project]
      summary: Run one of this project's jobs now
      description: |
        Addressed **by code, not by id**. A code is what the calling project
        knows about itself, and it also means the lookup is naturally scoped to
        the project rather than scoped by a check somebody could forget to write.

        It queues a row and hands it over; it does not execute anything inline.
        Both halves matter: the row means the request is durable and appears in
        the history like any other run, and the hand-off means it obeys the same
        single-run rule and concurrency cap as a scheduled one.
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
          example: daily-report
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          run_id: { type: integer, format: int64 }
                          job_id: { type: integer, format: int64 }
                          code: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No job with that code in this project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /runs:
    get:
      tags: [Project]
      summary: This project's run history
      description: Defaults to the last 24 hours when no `from` is given.
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/RunStatus" } }
        - { name: from, in: query, schema: { type: string, format: date }, description: Inclusive. }
        - { name: to, in: query, schema: { type: string, format: date }, description: Inclusive. }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: The run list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total: { type: integer }
                          runs:
                            type: array
                            items: { $ref: "#/components/schemas/RunRow" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /admin/jobs/{id}/run:
    post:
      tags: [Operator]
      summary: Run a job now
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          run_id: { type: integer, format: int64 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/runs:
    get:
      tags: [Operator]
      summary: Runs across the visible jobs
      description: |
        Honours the field restriction on `runs.read`. A role that may not see run
        output gets the rows with `output`, `error` or `request_url` empty rather
        than being refused the endpoint.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: job_id, in: query, schema: { type: integer, format: int64 } }
        - { name: project_id, in: query, schema: { type: integer, format: int64 } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/RunStatus" } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: The run list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total: { type: integer }
                          runs:
                            type: array
                            items: { $ref: "#/components/schemas/RunRow" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    projectKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A project's key. 24 random bytes, shown once when issued, stored only as
        a SHA-256 hash and compared in constant time. It reaches that project and
        nothing else. Rotating it invalidates the previous key immediately.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An operator token from `POST /login`. Valid 24 hours.
  responses:
    Unauthorized:
      description: No credential, or one that does not verify
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    Forbidden:
      description: Authenticated, but not allowed to do this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    NotFound:
      description: |
        No such thing, **or not yours**. The two are the same answer on purpose:
        telling somebody a job exists but belongs to a project they cannot see
        confirms an id, and an id is the only thing needed to probe the rest.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
...
  schemas:
    Envelope:
      type: object
      description: The shape of every response.
      properties:
        status: { type: boolean }
        message: { type: string }
        data: {}
        errors: {}
    RunStatus:
      type: string
      description: |
        `timeout` is not `failed`. When the client gives up the target may still
        be working, so a timeout triggers no chain step and is never retried.
        `skipped` means it did not run: either the previous run was still going
        and the job is `single_run`, or a chain step was blocked.
      enum: [pending, running, success, failed, timeout, skipped]
    RunTrigger:
      type: string
      enum: [schedule, manual, chain, api]
...
    Run:
      type: object
      properties:
        id: { type: integer, format: int64 }
        job_id: { type: integer, format: int64 }
        planned_minute: { type: string, format: date-time, nullable: true }
        trigger: { $ref: "#/components/schemas/RunTrigger" }
        parent_run_id: { type: integer, format: int64, nullable: true, description: Set when this run is a chain step. }
        status: { $ref: "#/components/schemas/RunStatus" }
        attempt: { type: integer }
        instance_id: { type: string, description: Which replica claimed it. }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        duration_ms: { type: integer, nullable: true }
        http_status:
          type: integer
          nullable: true
          description: Absent when no response arrived at all. "No answer" and "the server said 0" are not the same thing.
        request_url: { type: string }
        output: { type: string, description: Up to 16000 characters of the response body. }
        error: { type: string }
        created_at: { type: string, format: date-time }
    RunRow:
      allOf:
        - $ref: "#/components/schemas/Run"
        - type: object
          properties:
            job_code: { type: string }
            job_name: { type: string }
            project_slug: { type: string }
            project_name: { type: string }
...